Split PatternBuilder object out of QueryObjectBuilder
To make the object smaller and more understandable. It's just moving
around.

Bug: T272980
