+++ src/sparql/QueryObjectBuilder.ts
@@ -1,14 +1,12 @@
 import rdfNamespaces from '@/sparql/rdfNamespaces';
 import QueryRepresentation, { Condition } from '@/sparql/QueryRepresentation';
-import { FilterPattern, Pattern, SelectQuery, Triple } from 'sparqljs';
-import PropertyValueRelation from '@/data-model/PropertyValueRelation';
-import TripleBuilder from '@/sparql/TripleBuilder';
+import { Pattern, SelectQuery } from 'sparqljs';
 import ConditionRelation from '@/data-model/ConditionRelation';
-import ReferenceRelation from '@/data-model/ReferenceRelation';
+import PatternBuilder from '@/sparql/PatternBuilder';
 
 export default class QueryObjectBuilder {
 	private queryObject: SelectQuery;
-	private tripleBuilder: TripleBuilder;
+	private patternBuilder: PatternBuilder;
 
 	public constructor() {
 		this.queryObject = {
@@ -19,7 +17,7 @@ export default class QueryObjectBuilder {
 			type: 'query',
 			prefixes: rdfNamespaces,
 		};
-		this.tripleBuilder = new TripleBuilder();
+		this.patternBuilder = new PatternBuilder();
 	}
 
 	public buildFromQueryRepresentation( queryRepresentation: QueryRepresentation ): SelectQuery {
@@ -81,12 +79,7 @@ export default class QueryObjectBuilder {
 			}
 
 			if ( isOnlyNegateQuery ) {
-				this.queryObject.where.unshift(
-					{
-						type: 'bgp',
-						triples: [ this.tripleBuilder.buildAnyValueTripe() ],
-					},
-				);
+				this.queryObject.where.unshift( this.patternBuilder.buildAnyValuePattern() );
 			}
 		}
 
@@ -104,7 +97,7 @@ export default class QueryObjectBuilder {
 	private buildUnion( conditions: Condition[] ): void {
 		const unionConditions = [];
 		for ( let i = 0; i < conditions.length; i++ ) {
-			unionConditions.push( ...this.buildIndividualCondition( conditions[ i ], i ) );
+			unionConditions.push( ...this.patternBuilder.buildValuePatternFromCondition( conditions[ i ], i ) );
 		}
 		const union: Pattern = {
 			type: 'union',
@@ -116,96 +109,13 @@ export default class QueryObjectBuilder {
 		this.queryObject.where.push( union );
 	}
 
-	private buildIndividualCondition( condition: Condition, conditionIndex: number ): Pattern[] {
-		const negate: boolean = condition.negate || false;
-
-		const referenceTriple: Triple = {
-			subject: {
-				termType: 'Variable',
-				value: 'item',
-			},
-			predicate: {
-				termType: 'NamedNode',
-				value: rdfNamespaces.p + condition.propertyId,
-			},
-			object: {
-				termType: 'Variable',
-				value: 'statement' + conditionIndex,
-			},
-		};
-
-		const bgp: Pattern = {
-			type: 'bgp',
-			triples: [ this.tripleBuilder.buildTripleFromQueryCondition( condition, conditionIndex ) ],
-		};
-
-		const bgpReference: Pattern = {
-			type: 'bgp',
-			triples: [ referenceTriple ],
-		};
-
-		let referenceExistsOrNot = '';
-
-		if ( condition.referenceRelation === ReferenceRelation.With ) {
-			referenceExistsOrNot = 'exists';
-		} else if ( condition.referenceRelation === ReferenceRelation.Without ) {
-			referenceExistsOrNot = 'notexists';
-		}
-
-		const referenceFilter: FilterPattern = {
-			type: 'filter',
-			expression: {
-				type: 'operation',
-				operator: referenceExistsOrNot,
-				args: [
-					{
-						type: 'bgp',
-						triples: [
-							this.tripleBuilder.buildReferenceFilterTriple( condition, conditionIndex ),
-						],
-					},
-				],
-			},
-		};
-
-		const patterns: Pattern[] = [];
-
-		if ( negate === true ) {
-			patterns.push( {
-				type: 'minus',
-				patterns: [ bgp ],
-			} );
-		} else {
-			if ( condition.referenceRelation === ReferenceRelation.With ||
-					condition.referenceRelation === ReferenceRelation.Without ) {
-				patterns.push( bgpReference );
-				patterns.push( bgp );
-				patterns.push( referenceFilter );
-			} else {
-				patterns.push( bgp );
-			}
-		}
-
-		if ( condition.propertyValueRelation === PropertyValueRelation.NotMatching ) {
-			const filterCondition = {
-				type: 'minus',
-				patterns: [ {
-					type: 'bgp',
-					triples: [ this.tripleBuilder.buildTripleForNotMatchingValue( condition ) ],
-				} ],
-			};
-
-			patterns.push( filterCondition as Pattern );
-		}
-
-		return patterns;
-	}
-
 	private buildFromQueryCondition( condition: Condition, conditionIndex: number ): void {
 		if ( !this.queryObject.where ) {
 			this.queryObject.where = [];
 		}
-		this.queryObject.where.push( ...this.buildIndividualCondition( condition, conditionIndex ) );
+		this.queryObject.where.push(
+			...this.patternBuilder.buildValuePatternFromCondition( condition, conditionIndex ),
+		);
 		return;
 	}
 
@@ -226,12 +136,7 @@ export default class QueryObjectBuilder {
 			where: [
 				{
 					type: 'service',
-					patterns: [
-						{
-							type: 'bgp',
-							triples: [ this.tripleBuilder.buildLabelServiceTriple() ],
-						},
-					],
+					patterns: [ this.patternBuilder.buildLabelServicePattern() ],
 					name: {
 						termType: 'NamedNode',
 						value: rdfNamespaces.wikibase + 'label',
